you can load multiple recordings now and it should work
