+++ freemocap_blender_addon/core_functions/create_rig/add_rig_method_enum.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class AddRigMethods(Enum):
+    RIGIFY = "rigify"
+    BY_BONE = "by_bone"
