+++ freemocap_blender_addon/core_functions/create_rig/add_rig_rigify_method.py
@@ -1,102 +1,30 @@
-from enum import Enum
 import math as m
-import re
-from typing import Dict, List, Optional
-from freemocap_blender_addon.data_models.armatures.armature_bone_info import ArmatureBoneInfo
-from freemocap_blender_addon.data_models.poses.pose_element import PoseElement
+from typing import Dict
+
 import bpy
-import mathutils
 import addon_utils
-
-from freemocap_blender_addon.system.constants import (
-    FREEMOCAP_ARMATURE,
-    UE_METAHUMAN_SIMPLE_ARMATURE,
-    FREEMOCAP_TPOSE,
-    FREEMOCAP_APOSE,
-    UE_METAHUMAN_DEFAULT,
-    UE_METAHUMAN_TPOSE,
-)
-from freemocap_blender_addon.data_models.bones.bone_constraints import (
-    ALL_BONES_CONSTRAINT_DEFINITIONS,
-    ConstraintType,
-    CopyLocationConstraint,
-    DampedTrackConstraint,
-    IKConstraint,
-    LimitRotationConstraint,
-    LockedTrackConstraint,
-)
-from freemocap_blender_addon.data_models.bones.ik_control_bones import (
-    ik_control_bones,
-)
-from freemocap_blender_addon.data_models.bones.ik_pole_bones import ik_pole_bones
-from freemocap_blender_addon.data_models.armatures.bone_name_map import (
-    bone_name_map,
-)
-from freemocap_blender_addon.data_models.data_references import ArmatureType, PoseType
+import mathutils
 
 
-class AddRigMethods(Enum):
-    RIGIFY = "rigify"
-    BY_BONE = "by_bone"
+def ensure_rigify() -> None:
+    _, rigify_enabled = addon_utils.check("rigify")
 
-def add_rig(
-    bone_data: Dict[str, Dict[str, float]],
-    rig_name: str,
-    parent_object: bpy.types.Object,
-    add_rig_method: AddRigMethods = AddRigMethods.BY_BONE,
-    keep_symmetry: bool = False,
-    add_fingers_constraints: bool = False,
-    use_limit_rotation: bool = False,
-) -> bpy.types.Object:
+    if not rigify_enabled:
+        try:
+            print("Rigify not found - enabling Rigify addon...")
+            addon_utils.enable(
+                "rigify", default_set=True, persistent=True, handle_error=print
+            )
+        except Exception as e:
+            print(f"Error enabling Rigify addon - \n\n{e}")
+            raise e
 
-    # Deselect all objects
-    for object in bpy.data.objects:
-        object.select_set(False)
+    _, rigify_enabled = addon_utils.check("rigify")
 
-    if add_rig_method == AddRigMethods.RIGIFY:
-        rig = add_rig_rigify(
-            bone_data=bone_data,
-            rig_name=rig_name,
-            parent_object=parent_object,
-            keep_symmetry=keep_symmetry,
+    if not rigify_enabled:
+        raise Exception(
+            "Rigify not enabled, please enable it in your blender preferences and then close blender before retrying"
         )
-    elif add_rig_method == AddRigMethods.BY_BONE:
-        rig = add_rig_by_bone(
-            bone_data=bone_data,
-            rig_name=rig_name,
-            armature=ArmatureType.FREEMOCAP,
-            pose=PoseType.FREEMOCAP_TPOSE,
-            add_ik_constraints=False,
-        )
-    else:
-        raise ValueError(f"Invalid add rig method: {add_rig_method}")
-
-    # Change mode to object mode
-    bpy.ops.object.mode_set(mode="OBJECT")
-
-    # TODO: make sure this still adds constraints properly
-    add_constraints(
-        rig=rig,
-        add_fingers_constraints=add_fingers_constraints,
-        parent_object=parent_object,
-        armature=ArmatureType.FREEMOCAP,
-        pose=PoseType.FREEMOCAP_TPOSE,
-        use_limit_rotation=use_limit_rotation,
-    )
-
-    ### Bake animation to the rig ###
-    # Get the empties ending frame
-    ending_frame = int(bpy.data.actions[0].frame_range[1])
-    # Bake animation
-    bpy.ops.nla.bake(frame_start=1, frame_end=ending_frame, bake_types={"POSE"})
-
-    # Change back to Object Mode
-    bpy.ops.object.mode_set(mode="OBJECT")
-
-    # Deselect all objects
-    bpy.ops.object.select_all(action="DESELECT")
-
-    return rig
 
 
 def add_rig_rigify(
@@ -898,316 +826,3 @@ def add_rig_rigify(
     bpy.ops.pose.armature_apply(selected=False)
 
     return rig
-
-
-def add_rig_by_bone(
-    bone_data: Dict[str, Dict[str, float]],
-    rig_name: str,
-    armature: Dict[str, ArmatureBoneInfo] = ArmatureType.FREEMOCAP,
-    pose: Dict[str, PoseElement] = PoseType.FREEMOCAP_TPOSE,
-    add_ik_constraints: bool = False,
-) -> bpy.types.Object:
-    print("Adding rig to scene bone by bone...")
-    if armature == ArmatureType.UE_METAHUMAN_SIMPLE:
-        armature_name = UE_METAHUMAN_SIMPLE_ARMATURE
-    elif armature == ArmatureType.FREEMOCAP:
-        armature_name = FREEMOCAP_ARMATURE
-    else:
-        raise ValueError("Invalid armature name")
-
-    # Get rig height as the sum of the major bones length in a standing position. Assume foot declination angle of 23º
-    avg_ankle_projection_length = (
-        m.sin(m.radians(23)) * bone_data["foot.R"]["median"]
-        + m.sin(m.radians(23)) * bone_data["foot.L"]["median"]
-    ) / 2
-    avg_shin_length = (
-        bone_data["shin.R"]["median"] + bone_data["shin.L"]["median"]
-    ) / 2
-    avg_thigh_length = (
-        bone_data["thigh.R"]["median"] + bone_data["thigh.L"]["median"]
-    ) / 2
-
-    # Add the armature
-    bpy.ops.object.armature_add(
-        enter_editmode=False,
-        align="WORLD",
-        location=(0, 0, 0),
-    )
-
-    # Rename the armature
-    bpy.data.armatures[0].name = "root"
-    # Get reference to armature
-    rig = bpy.data.objects["Armature"]
-    # Rename the rig object to pelvis
-    rig.name = "root"
-    # Get reference to the renamed armature
-    rig = bpy.data.objects["root"]
-
-    # Change to edit mode
-    bpy.ops.object.mode_set(mode="EDIT")
-
-    # Remove the default bone
-    rig.data.edit_bones.remove(rig.data.edit_bones["Bone"])
-
-    # Get the inverse bone_map_dict
-    inv_bone_name_map = {
-        value: key for key, value in bone_name_map[armature_name].items()
-    }
-
-    # Iterate over the armature dictionary
-    for bone in armature:
-
-        # Get the reference to the parent of the bone if its not root
-        parent_name = armature[bone].parent_bone
-        if parent_name != "root":
-            parent_bone = rig.data.edit_bones[parent_name]
-
-        # Add the new bone
-        rig_bone = rig.data.edit_bones.new(bone)
-
-        # Set the bone head position
-        if bone in ("pelvis"):
-            rig_bone.head = mathutils.Vector(
-                [
-                    0,
-                    0,
-                    avg_ankle_projection_length + avg_shin_length + avg_thigh_length,
-                ]
-            )
-        else:
-            # Set the bone position relative to its parent
-            if armature[bone].parent_position == "head":
-                rig_bone.head = parent_bone.head
-            elif armature[bone].parent_position == "tail":
-                rig_bone.head = parent_bone.tail
-
-        # Get the bone vector
-        if inv_bone_name_map[bone] not in bone_data:
-            bone_vector = mathutils.Vector(
-                [0, 0, armature[bone].default_length]
-            )
-        else:
-            bone_vector = mathutils.Vector(
-                [0, 0, bone_data[inv_bone_name_map[bone]]["median"]]
-            )
-
-        # Get the rotation matrix
-        rotation_matrix = mathutils.Euler(
-            mathutils.Vector(pose[bone].rotation),
-            "XYZ",
-        ).to_matrix()
-
-        # Rotate the bone vector
-        rig_bone.tail = rig_bone.head + rotation_matrix @ bone_vector
-
-        # Assign the roll to the bone
-        rig_bone.roll = pose[bone].roll
-
-        # Parent the bone if its parent exists
-        if parent_name != "root":
-            rig_bone.parent = parent_bone
-            rig_bone.use_connect = armature[bone].connected
-
-    # Special armature conditions
-    if armature_name == UE_METAHUMAN_SIMPLE_ARMATURE:
-        # Change parents of thigh bones
-        rig.data.edit_bones["thigh_r"].use_connect = False
-        rig.data.edit_bones["thigh_l"].use_connect = False
-        rig.data.edit_bones["thigh_r"].parent = rig.data.edit_bones["pelvis"]
-        rig.data.edit_bones["thigh_l"].parent = rig.data.edit_bones["pelvis"]
-
-    # Add the ik bones if specified
-    if add_ik_constraints:
-        for ik_control in ik_control_bones:
-            ik_bone = rig.data.edit_bones.new(ik_control)
-            ik_bone.head = rig.data.edit_bones[
-                bone_name_map[armature_name][
-                    ik_control_bones[ik_control].controlled_bone
-                ]
-            ].head
-            ik_bone.tail = ik_bone.head + mathutils.Vector(
-                ik_control_bones[ik_control].tail_relative_position
-            )
-        for ik_pole in ik_pole_bones:
-            ik_bone = rig.data.edit_bones.new(ik_pole)
-            ik_bone.head = ik_pole_bones[ik_pole].head_position
-            ik_bone.tail = ik_pole_bones[ik_pole].tail_position
-
-    return rig
-
-
-def add_constraints(
-    rig: bpy.types.Object,
-    add_fingers_constraints: bool,
-    parent_object: bpy.types.Object,
-    armature: Dict[str, ArmatureBoneInfo] = ArmatureType.FREEMOCAP,
-    pose: Dict[str, PoseElement] = PoseType.FREEMOCAP_TPOSE,
-    use_limit_rotation: bool = False,
-) -> None:
-    if armature == ArmatureType.UE_METAHUMAN_SIMPLE:
-        armature_name = UE_METAHUMAN_SIMPLE_ARMATURE
-    elif armature == ArmatureType.FREEMOCAP:
-        armature_name = FREEMOCAP_ARMATURE
-    else:
-        raise ValueError("Invalid armature name")
-    
-    if pose == PoseType.FREEMOCAP_TPOSE:
-        pose_name = FREEMOCAP_TPOSE
-    elif pose == PoseType.FREEMOCAP_APOSE:
-        pose_name = FREEMOCAP_APOSE
-    elif pose == PoseType.UE_METAHUMAN_DEFAULT:
-        pose_name = UE_METAHUMAN_DEFAULT
-    elif pose == PoseType.UE_METAHUMAN_TPOSE:
-        pose_name = UE_METAHUMAN_TPOSE
-    else:
-        raise ValueError("Invalid pose name")
-
-    print("Adding bone constraints...")
-    # TODO: getting key error in this function with Failed to add rig: 'bpy_prop_collection[key]: key "pelvis.R" not found'
-    # Change to pose mode
-    bpy.context.view_layer.objects.active = rig
-    bpy.ops.object.mode_set(mode="POSE")
-    #
-    # # Define the hand bones damped track target as the hand middle empty if they were already added
-    # try:
-    #     right_hand_middle_name = bpy.data.objects['right_hand_middle'].name
-    #     # Right Hand Middle Empty exists. Use hand middle as target
-    #     hand_damped_track_target = 'hand_middle'
-    # except:
-    #     # Hand middle empties do not exist. Use hand_index as target
-    #     hand_damped_track_target = 'index'
-
-    # Create each constraint
-    for (
-        bone_name,
-        constraint_definitions,
-    ) in ALL_BONES_CONSTRAINT_DEFINITIONS.items():
-        # If pose bone does not exist, skip it
-        if bone_name not in rig.pose.bones:
-            continue
-        
-        if not isinstance(constraint_definitions, list):
-            raise Exception(f"Constraint definitions for {bone_name} must be a list")
-
-        # If it is a finger bone amd add_fingers_constraints is False continue with the next bone
-        if (
-            not add_fingers_constraints
-            and len(
-                [
-                    finger_part
-                    for finger_part in [
-                        "palm",
-                        "thumb",
-                        "index",
-                        "middle",
-                        "ring",
-                        "pinky",
-                    ]
-                    if finger_part
-                    in constraint_definitions
-                ]
-            )
-            > 0
-        ):
-            continue
-
-        for constraint in constraint_definitions:
-            # if "target" in constraint.keys():
-            #     base_target_name = constraint["target"]
-            #     actual_target_name = get_actual_empty_target_name(empty_names = empty_names,
-            #                                                       base_target_name = base_target_name)
-            #     constraint["target"] = actual_target_name
-            appended_number_string = get_appended_number(parent_object.name)
-            if appended_number_string is not None:
-                if hasattr(constraint, "target"):
-                    constraint.target = constraint.target + appended_number_string
-            # Add new constraint determined by type
-            if not use_limit_rotation and constraint.type == ConstraintType.LIMIT_ROTATION:
-                continue
-            else:
-                try:
-                    bone_constraint = rig.pose.bones[bone_name_map[armature_name][bone_name]].constraints.new(
-                        constraint.type.value
-                    )
-                except:
-                    print(f"Failed to add rig: {bone_name} constraint {constraint.type}")
-                    continue
-
-                # Define aditional parameters based on the type of constraint
-            if isinstance(constraint, LimitRotationConstraint):
-                bone_constraint.use_limit_x = constraint.use_limit_x
-                bone_constraint.min_x = m.radians(constraint.min_x)
-                bone_constraint.max_x = m.radians(constraint.max_x)
-                bone_constraint.use_limit_y = constraint.use_limit_y
-                bone_constraint.min_y = m.radians(constraint.min_y)
-                bone_constraint.max_y = m.radians(constraint.max_y)
-                bone_constraint.use_limit_z = constraint.use_limit_z
-                bone_constraint.min_z = m.radians(constraint.min_z)
-                bone_constraint.max_z = m.radians(constraint.max_z)
-                bone_constraint.owner_space = constraint.owner_space.value
-            elif isinstance(constraint, CopyLocationConstraint):
-                bone_constraint.target = bpy.data.objects[constraint.target]
-            elif isinstance(constraint, LockedTrackConstraint):
-                bone_constraint.target = bpy.data.objects[constraint.target]
-                bone_constraint.track_axis = constraint.track_axis[pose_name].value
-                bone_constraint.lock_axis = constraint.lock_axis[pose_name].value
-                bone_constraint.influence = constraint.influence
-            elif isinstance(constraint, DampedTrackConstraint):
-                bone_constraint.target = bpy.data.objects[constraint.target]
-                bone_constraint.track_axis = constraint.track_axis.value
-            elif isinstance(constraint, IKConstraint):
-                bone_constraint.target = bpy.data.objects[constraint.target]
-                bone_constraint.pole_target = bpy.data.objects[
-                    constraint.pole_target
-                ]
-                bone_constraint.chain_count = constraint.chain_count
-                bone_constraint.pole_angle = constraint.pole_angle
-
-
-def ensure_rigify() -> None:
-    _, rigify_enabled = addon_utils.check("rigify")
-
-    if not rigify_enabled:
-        try:
-            print("Rigify not found - enabling Rigify addon...")
-            addon_utils.enable(
-                "rigify", default_set=True, persistent=True, handle_error=print
-            )
-        except Exception as e:
-            print(f"Error enabling Rigify addon - \n\n{e}")
-            raise e
-
-    _, rigify_enabled = addon_utils.check("rigify")
-
-    if not rigify_enabled:
-        raise Exception(
-            "Rigify not enabled, please enable it in your blender preferences and then close blender before retrying"
-        )
-
-
-def get_appended_number(rig_name: str) -> Optional[str]:
-    pattern = r"\.0[0-9]{2}$"
-    match = re.search(pattern, rig_name)
-    return match.group() if match else None
-
-
-def get_actual_empty_target_name(empty_names: List[str], base_target_name: str) -> str:
-    """
-    Get the actual empty target name based on the constraint target name,
-    this is mostly to give us the ability to load multiple recorings, because
-    blender will append `.001`, `.002`  the names of emtpies of the 2nd, 3rd, etc to avoid name collisions
-
-    So basically, if the base_target name is `hips_center` this will look for empties named `hips_center`,
-      `hips_center.001`, `hips_center.002`, etc in the provided `empty_names` list and return that
-    """
-
-    actual_target_name = None
-    for empty_name in empty_names:
-        if base_target_name in empty_name:
-            actual_target_name = empty_name
-            break
-
-    if actual_target_name is None:
-        raise ValueError(f"Could not find empty target for {base_target_name}")
-
-    return actual_target_name
