+++ freemocap_blender_addon/core_functions/create_rig/save_bone_and_joint_angles_from_rig.py
@@ -3,7 +3,8 @@ from pathlib import Path
 from typing import Dict
 
 import bpy
-from freemocap_blender_addon.data_models.bones.bone_constraints import ALL_BONES_CONSTRAINT_DEFINITIONS
+
+from freemocap_blender_addon.data_models.bones.bone_constraints import get_bone_constraint_definitions
 
 
 def save_bone_and_joint_angles_from_rig(rig: bpy.types.Object,
@@ -21,7 +22,7 @@ def save_bone_and_joint_angles_from_rig(rig: bpy.types.Object,
         frame_data = {}
         all_bone_data[frame_number] = frame_data
         for bone in rig.pose.bones:
-            if bone.name not in ALL_BONES_CONSTRAINT_DEFINITIONS.keys():
+            if bone.name not in get_bone_constraint_definitions().keys():
                 continue
             frame_data[bone.name] = get_bone_data(bone)
