+++ freemocap_blender_addon/core_functions/data_models/armatures/root_bone_name.py
@@ -1,0 +1,1 @@
+ROOT_BONE_NAME = "root_bone"
